# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GmbH API
  version: 1.2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/items:
    get:
      tags:
        - Item
      summary: Request all items
      responses:
        '200':
          description: Items list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Item
      summary: Create one item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/items/{id}:
    get:
      tags:
        - Item
      summary: Request item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Item
      summary: Update item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdateRequest'
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Item
      summary: Delete item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/events:
    get:
      tags:
        - Event
      summary: Request all events
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Event
      summary: Create one event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/{id}:
    put:
      tags:
        - Event
      summary: Update one event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateRequest'
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Event
      summary: Delete one event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        errors:
          type: object
          properties:
            msg:
              type: string
          required:
            - msg
      required:
        - errors
...
    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: number
        price:
          type: number
        tax:
          type: number
        sort:
          type: number
        categoryId:
          type: string
        unitId:
          type: string
        enabled:
          type: boolean
    ItemResponse:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
      required:
        - items
    ItemCreateRequest:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
    ItemUpdateRequest:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
...
    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        beginDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventResponse:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/Event'
      required:
        - event
    EventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        activeEventId:
          type: string
          nullable: true
      required:
        - events
        - activeEventId
    EventCreateRequest:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/Event'
        importFromEventId:
          type: string
          nullable: true
        include:
          type: object
          properties:
            units:
              type: boolean
            categories:
              type: boolean
            items:
              type: boolean
            areas:
              type: boolean
            tables:
              type: boolean
      required:
        - event
    EventUpdateRequest:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/Event'
      required:
        - event
...